feat: Create bae routes navigation.
