+++ src/app/centrals/create/page.tsx
@@ -1,0 +1,8 @@
+import { CreateCentralPage } from "@pages/centrals/create";
+
+export default function CreateCentral() {
+
+  return (
+    <CreateCentralPage />
+  );
+}
