+++ src/app/centrals/page.tsx
@@ -1,0 +1,7 @@
+import { CentralsPage } from "@pages/centrals";
+
+export default function Centrals() {
+  return (
+    <CentralsPage />
+  );
+}
