+++ src/app/centrals/styles.css.ts
@@ -1,0 +1,3 @@
+import { style } from "@vanilla-extract/css";
+
+export const centralsLayoutStyle = style({});
