+++ src/app/layout.tsx
@@ -3,6 +3,12 @@ import { ReactNode } from "react";
 
 import { fonts } from "@components/styles/fonts";
 import "@styles/global.css";
+import { TanstackProvider } from "@main/providers/tanstack-provider";
+import { Sidebar } from "@components/core/sidebar";
+import { MENU_RESOURCES_CONFIGS } from "@config/menu";
+import * as styles from "./styles.css";
+import { ClientHeader } from "@components/core/header/client-header";
+  import { ToastContainer } from 'react-toastify';
 
 export const metadata: Metadata = {
   title: "Defense IA | Middlewares e Centrais",
@@ -10,8 +16,21 @@ export const metadata: Metadata = {
 
 export default function RootLayout({ children }: { children: ReactNode }) {
   return (
-    <html lang="pt-br" className={fonts.nunito}>
-      <body>{children}</body>
+    <html lang="pt-br">
+      <body className={fonts.nunito}>
+        <TanstackProvider>
+          <ToastContainer />
+            <main className={styles.mainLayoutStyle}>
+              <Sidebar.Root>
+                <Sidebar.Menu resources={MENU_RESOURCES_CONFIGS} />
+              </Sidebar.Root>
+              <div className={styles.containerPageStyles}>
+                <ClientHeader />
+                {children}
+              </div>
+            </main>
+        </TanstackProvider>
+      </body>
     </html>
   );
 }
